Initial version of the app. Everything is working although still needs some tweaks.
That's my first android app and I hacked it in a weekend so still lots of improvements.
I've tried to have as less dependencies as possible, that's whey some pieces of the code would be better done with guava or a DI framework.
Need to add tests as I didn't have enough time to play with them and, ideally, do some refactor with that.