Added limitations + bug fixes
Added the possibility to define how the urticaria has limited you (work, socialize...) in a similar way as angioedema  was defined.
The db has been updated to have this field and also to support pictures as well (although it's still not implemented).
Fixed a small bug with export, now it should work now in older versions as we are not using Environment.DIRECTORY_DOCUMENTS.